Initial CMR integration
