+++ src/main/java/gov/nasa/podaac/swodlr/l2rasterproduct/L2RasterProductService.java
@@ -1,5 +1,6 @@
 package gov.nasa.podaac.swodlr.l2rasterproduct;
 
+import gov.nasa.podaac.swodlr.cmr.SwotCmrLookupService;
 import gov.nasa.podaac.swodlr.exception.SwodlrException;
 import gov.nasa.podaac.swodlr.producthistory.ProductHistory;
 import gov.nasa.podaac.swodlr.producthistory.ProductHistoryRepository;
@@ -9,10 +10,13 @@ import gov.nasa.podaac.swodlr.status.State;
 import gov.nasa.podaac.swodlr.status.Status;
 import gov.nasa.podaac.swodlr.status.StatusRepository;
 import gov.nasa.podaac.swodlr.user.User;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.UUID;
 import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import reactor.core.publisher.Mono;
 
 @Service
 public class L2RasterProductService {
@@ -28,23 +32,48 @@ public class L2RasterProductService {
   @Autowired
   ProductHistoryRepository productHistoryRepository;
 
+  @Autowired
+  SwotCmrLookupService swotCmrLookupService;
+
   @Transactional
-  public L2RasterProduct createL2RasterProduct(User user, UUID definitionId,
-      int cycle, int pass, int scene) {
-    var queryResult = rasterDefinitionRepository.findById(definitionId);
-    if (!queryResult.isPresent()) {
-      throw new SwodlrException("Definition not found");
-    }
+  public Mono<L2RasterProduct> createL2RasterProduct(
+      User user,
+      UUID definitionId,
+      int cycle,
+      int pass,
+      int scene
+  ) {
+    return Mono
+      .defer(() -> {
+        var result = rasterDefinitionRepository.findById(definitionId);
+        if (!result.isPresent()) {
+          return Mono.error(new SwodlrException("Definition not found"));
+        }
 
-    RasterDefinition definition = queryResult.get();
-    L2RasterProduct product = new L2RasterProduct(definition, cycle, pass, scene);
-    Status status = new Status(product, State.NEW);
-    ProductHistory history = new ProductHistory(user, product);
+        Map<String, Object> data = new HashMap<>();
+        data.put("definition", result.get());
 
-    product = l2RasterProductRepository.save(product);
-    statusRepository.save(status);
-    productHistoryRepository.save(history);
+        return Mono.just(data);
+      })
+      .flatMap((Map<String, Object> data) -> {
+        return swotCmrLookupService.findGranules(cycle, pass, scene)
+            .map((granules) -> {
+              data.put("granules", granules);
+              return data;
+            });
+      })
+      .map((Map<String, Object> data) -> {
+        RasterDefinition definition = (RasterDefinition) data.get("definition");
 
-    return product;
+        L2RasterProduct product = new L2RasterProduct(definition, cycle, pass, scene);
+        Status status = new Status(product, State.NEW);
+        ProductHistory history = new ProductHistory(user, product);
+
+        product = l2RasterProductRepository.save(product);
+        statusRepository.save(status);
+        productHistoryRepository.save(history);
+
+        return product;
+      });
   }
 }
