+++ src/main/resources/graphql-documents/query/granule_lookup.graphql
